back/requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
dill==0.2.5
Django==2.1.7
future==0.17.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
PyFunctional==1.2.0
pylint==2.2.2
python-resize-image==1.1.18
pytz==2018.9
requests==2.21.0
six==1.12.0
tabulate==0.8.3
urllib3==1.24.1
wrapt==1.11.1
